Move game end post to game official event
